requirements.txt
Flask==3.0.2
peewee==3.17.6
python-dotenv==1.0.1